Extract notification handling to separate class
